+++ packages/bluesky/CHANGELOG.md
@@ -7,6 +7,7 @@
   - `.pushBlob`
   - `.transferAccount`
 - Added `did` property on `ServerInfo` object. Returned from `ServerService.describeServer`. ([#1279](https://github.com/myConsciousness/atproto.dart/issues/1279))
+- Added `.getServiceAuth` on `ServerService`. ([#1280](https://github.com/myConsciousness/atproto.dart/issues/1280))
 
 ## v0.15.8